3949 - Add TablePaginateds & UserFilters
