+++ src/meta/api/request/admin/users.ts
@@ -1,8 +1,5 @@
 import { Request } from 'express'
 
-import { CountryIso } from 'meta/area'
-import { RoleName } from 'meta/user'
-
 export type UsersRequest<QueryParams = any, Body = any> = Request<
   never,
   never,
@@ -17,10 +14,3 @@ export type UsersRequest<QueryParams = any, Body = any> = Request<
     orderByDirection?: string
   }
 >
-
-export type DecodedUserFilters = {
-  administrators?: boolean
-  countries?: Array<CountryIso>
-  fullName?: string
-  roles?: Array<RoleName>
-}
