+++ src/meta/tablePaginated/index.ts
@@ -1,20 +1,8 @@
-export enum TablePaginatedOrderByDirection {
-  asc = 'asc',
-  desc = 'desc',
-}
-
-export type TablePaginatedOrderBy = {
-  direction?: TablePaginatedOrderByDirection
-  property?: string
-}
-
-export type TablePaginatedCount = {
-  total: number
-}
-
-export enum TablePaginatedFilterType {
-  TEXT = 'text',
-  SWITCH = 'switch',
-}
-
-export type TablePaginatedFilterValues = boolean | string | Array<string>
+export type {
+  TablePaginatedCount,
+  TablePaginatedFilterValues,
+  TablePaginatedOrderBy,
+  UserFilters,
+} from './tablePaginated'
+export { TablePaginatedFilterType, TablePaginatedOrderByDirection } from './tablePaginated'
+export { TablePaginateds } from './tablePaginateds'
